backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
faker==18.13.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.4.4
typing_extensions==4.7.1
